+++ database/migrations/2020_12_14_051822_create_order_product_table.php
@@ -15,8 +15,8 @@ class CreateOrderProductTable extends Migration
     {
         Schema::create('order_product', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('order_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->integer('quantity')->unsigned()->default(0);
             $table->integer('sub_total')->unsigned()->default(0);
             $table->timestamps();
